Divide components into separate files
