+++ src/components/stages/Stage5/GameInstructions.tsx
@@ -1,0 +1,10 @@
+export function GameInstructions({ startGame }: { startGame: () => void }) {
+  // TODO
+  return (
+    <>
+      <h2>Zombie cookie game</h2>
+      <p>Instructions go here&hellip;</p>
+      <button onClick={startGame}>Start game</button>
+    </>
+  );
+}
