+++ src/components/stages/Stage5/Score.tsx
@@ -1,0 +1,5 @@
+export function Score({ points }: { points: number }) {
+  return (
+    <p className="score">Score: {points}</p>
+  );
+}
